Issue with env
There seems to be a problem with both the model and the env. Playing as
white (via stonefish.eval), the model will defeat random almost every time (95%
with the remaining 5% being draws). Playing as black it will lose every
time. This is very concerning, and must be a bug somewhere.

The env object comes with a "reset" option that makes it so the board is always
configured so white is to move. Setting this allows the model to achieve a win
rate close to the "white" case above. However, this only works with a
FastChessEnv running with a single environment. Running multiple single
environment versions in parallel works (see FakeEnv in rl_loop.py).